requirements.txt
streamlit
pillow
openpyxl
pytesseract
google-genai
